pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stretch3_zmq"]

[project]
name = "stretch3-zmq"
version = "0.1.0"
requires-python = ">=3.12, <3.13"
dependencies = [
    "elevenlabs>=2.32.0",
    "pyzmq>=27.1.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "dotenv>=0.9.9",
    "rich>=13.0.0",
    "scipy>=1.17.0",
    "hello-robot-stretch-body==0.7.29",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "pydantic==2.10.6",
    "ikpy>=3.4.2",
    "graphviz>=0.21",
    "urchin>=0.0.30",
    "msgpack>=1.1.2",
    "websockets==12.0",
    "opencv-python>=4.13.0.92",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
stretch3-zmq-driver = "stretch3_zmq.driver.__main__:main"

[dependency-groups]
dev = [
    "prek>=0.3.3",
    "pytest>=9.0.2",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.isort]
known-first-party = ["stretch3_zmq"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["stretch_body.*", "pyrealsense2.*", "cv2.*", "sounddevice.*", "msgpack.*", "scipy.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
